Crypto and small fixes

+++ modals/JET.js
@@ -1,5 +1,14 @@
 "use strict";
 
+// Crypton dokumentaatio suosittelee varmistamaan, että käytössä oleva Node.js -ympäristö sisältää crypto-moduulin.
+let crypto;
+try{
+    crypto = require("crypto");
+}
+catch(error){
+    throw "Crypto-moduuli ei ole käytettävissä. Lopetetaan."
+}
+
 const fs = require("fs");
 
 const datafile = "./data.json";
@@ -10,7 +19,7 @@ if(!fs.existsSync(datafile)){
         users: [
             {
                 username: "admin",
-                password: "kissakala"
+                password: "1a8c5de43cd21c6e4a0ea70d01bbb380660cd3192e4eb1dd2f24d79e12bce1e2" // kissakala
             }
         ],
         reservations: [
@@ -40,15 +49,25 @@ module.exports = {
             fs.readFile(datafile,"UTF-8",(error,data) => {
                 if(!error){
                     let currentData = JSON.parse(data);
-                    currentData.reservations.push(thisReservation);
-                    fs.writeFile(datafile,JSON.stringify(currentData,null,2),(error) => {
-                        if(!error){
-                            resolve();
-                        }
-                        else{
-                            reject(error);
-                        }
+                    
+                    let doesExist = currentData.reservations.find((reservation) => {
+                        return reservation.time.weekNumber === thisReservation.time.weekNumber && reservation.time.year === thisReservation.time.year && reservation.time.cellId === thisReservation.time.cellId;
                     });
+
+                    if(!doesExist){
+                        currentData.reservations.push(thisReservation);
+                        fs.writeFile(datafile,JSON.stringify(currentData,null,2),(error) => {
+                            if(!error){
+                                resolve();
+                            }
+                            else{
+                                reject(error);
+                            }
+                        });
+                    }
+                    else{
+                        reject();
+                    }
                 }
                 else{
                     reject(error);
@@ -60,7 +79,7 @@ module.exports = {
         return new Promise((resolve,reject) => {
             fs.readFile(datafile,"UTF-8",(error,data) => {
                 if(!error){
-                    resolve(JSON.parse(data).reservations);
+                    reject(JSON.parse(data).reservations);
                 }
                 else{
                     reject(error);
@@ -73,10 +92,12 @@ module.exports = {
             fs.readFile(datafile,"UTF-8",(error,data) => {
                 if(!error){
                     let users = JSON.parse(data).users;
+                    let hash = crypto.createHash("sha256").update(body.password).digest("hex");
+                    console.log(hash);
                     let thisUser = users.find((user) => {
                         return body.username === user.username;
                     });
-                    if(thisUser && body.password === thisUser.password){
+                    if(thisUser && hash === thisUser.password){
                         resolve();
                     }
                     else{
